requirements.txt
torch>=2.0
torchvision>=0.15
transformers>=4.40
numpy>=1.24
pandas>=2.0
scikit-learn>=1.4
matplotlib>=3.8
seaborn>=0.13
tqdm>=4.65
Pillow>=10.0
streamlit>=1.36
